@@ -1,10 +1,10 @@
 name: demo
 description: AutoSizeText Demo App
 
-version: 1.0.0+1
+version: 1.0.0+2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0" 
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
     path: ../
 
   bottom_navy_bar: ^4.2.0
-  material_design_icons_flutter: ^3.2.3695
+  material_design_icons_flutter: ^4.0.5755
 
 dev_dependencies:
   flutter_test: 
@@ -22,3 +22,5 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+
+publish_to: none
