@@ -1,10 +1,10 @@
 name: example
-description: AutoSizeText Example
+description: AutoSizeText example
 
 version: 1.0.0
 
 environment:
-sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
@@ -19,3 +19,5 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true 
+
+publish_to: none
