@@ -1,6 +1,7 @@
 part of auto_size_text;
 
-/// Flutter widget that automatically resizes text to fit perfectly within its bounds.
+/// Flutter widget that automatically resizes text to fit perfectly within its
+/// bounds.
 ///
 /// All size constraints as well as maxLines are taken into account. If the text
 /// overflows anyway, you should check if the parent widget actually constraints
@@ -11,8 +12,8 @@ class AutoSizeText extends StatefulWidget {
   /// If the [style] argument is null, the text will use the style from the
   /// closest enclosing [DefaultTextStyle].
   const AutoSizeText(
-    this.data, {
-    Key key,
+    String this.data, {
+    Key? key,
     this.textKey,
     this.style,
     this.strutStyle,
@@ -31,15 +32,13 @@ class AutoSizeText extends StatefulWidget {
     this.textScaleFactor,
     this.maxLines,
     this.semanticsLabel,
-  })  : assert(data != null,
-            'A non-null String must be provided to a AutoSizeText widget.'),
-        textSpan = null,
+  })  : textSpan = null,
         super(key: key);
 
   /// Creates a [AutoSizeText] widget with a [TextSpan].
   const AutoSizeText.rich(
-    this.textSpan, {
-    Key key,
+    TextSpan this.textSpan, {
+    Key? key,
     this.textKey,
     this.style,
     this.strutStyle,
@@ -58,47 +57,45 @@ class AutoSizeText extends StatefulWidget {
     this.textScaleFactor,
     this.maxLines,
     this.semanticsLabel,
-  })  : assert(textSpan != null,
-            'A non-null TextSpan must be provided to a AutoSizeText.rich widget.'),
-        data = null,
+  })  : data = null,
         super(key: key);
 
   /// Sets the key for the resulting [Text] widget.
   ///
   /// This allows you to find the actual `Text` widget built by `AutoSizeText`.
-  final Key textKey;
+  final Key? textKey;
 
   /// The text to display.
   ///
   /// This will be null if a [textSpan] is provided instead.
-  final String data;
+  final String? data;
 
   /// The text to display as a [TextSpan].
   ///
   /// This will be null if [data] is provided instead.
-  final TextSpan textSpan;
+  final TextSpan? textSpan;
 
   /// If non-null, the style to use for this text.
   ///
-  /// If the style's 'inherit' property is true, the style will be merged with
+  /// If the style's "inherit" property is true, the style will be merged with
   /// the closest enclosing [DefaultTextStyle]. Otherwise, the style will
   /// replace the closest enclosing [DefaultTextStyle].
-  final TextStyle style;
+  final TextStyle? style;
 
   // The default font size if none is specified.
-  static const double _defaultFontSize = 14.0;
+  static const double _defaultFontSize = 14;
 
   /// The strut style to use. Strut style defines the strut, which sets minimum
   /// vertical layout metrics.
   ///
   /// Omitting or providing null will disable strut.
   ///
-  /// Omitting or providing null for any properties of [StrutStyle] will result in
-  /// default values being used. It is highly recommended to at least specify a
-  /// font size.
+  /// Omitting or providing null for any properties of [StrutStyle] will result
+  /// in default values being used. It is highly recommended to at least specify
+  /// a font size.
   ///
   /// See [StrutStyle] for details.
-  final StrutStyle strutStyle;
+  final StrutStyle? strutStyle;
 
   /// The minimum text size constraint to be used when auto-sizing text.
   ///
@@ -124,17 +121,17 @@ class AutoSizeText extends StatefulWidget {
   /// Predefines all the possible font sizes.
   ///
   /// **Important:** PresetFontSizes have to be in descending order.
-  final List<double> presetFontSizes;
+  final List<double>? presetFontSizes;
 
   /// Synchronizes the size of multiple [AutoSizeText]s.
   ///
   /// If you want multiple [AutoSizeText]s to have the same text size, give all
   /// of them the same [AutoSizeGroup] instance. All of them will have the
   /// size of the smallest [AutoSizeText]
-  final AutoSizeGroup group;
+  final AutoSizeGroup? group;
 
   /// How the text should be aligned horizontally.
-  final TextAlign textAlign;
+  final TextAlign? textAlign;
 
   /// The directionality of the text.
   ///
@@ -149,20 +146,20 @@ class AutoSizeText extends StatefulWidget {
   /// its left.
   ///
   /// Defaults to the ambient [Directionality], if any.
-  final TextDirection textDirection;
+  final TextDirection? textDirection;
 
   /// Used to select a font when the same Unicode character can
   /// be rendered differently, depending on the locale.
   ///
   /// It's rarely necessary to set this property. By default its value
   /// is inherited from the enclosing app with `Localizations.localeOf(context)`.
-  final Locale locale;
+  final Locale? locale;
 
   /// Whether the text should break at soft line breaks.
   ///
   /// If false, the glyphs in the text will be positioned as if there was
   /// unlimited horizontal space.
-  final bool softWrap;
+  final bool? softWrap;
 
   /// Whether words which don't fit in one line should be wrapped.
   ///
@@ -171,11 +168,13 @@ class AutoSizeText extends StatefulWidget {
   final bool wrapWords;
 
   /// How visual overflow should be handled.
-  final TextOverflow overflow;
+  ///
+  /// Defaults to retrieving the value from the nearest [DefaultTextStyle] ancestor.
+  final TextOverflow? overflow;
 
   /// If the text is overflowing and does not fit its bounds, this widget is
   /// displayed instead.
-  final Widget overflowReplacement;
+  final Widget? overflowReplacement;
 
   /// The number of font pixels for each logical pixel.
   ///
@@ -187,7 +186,7 @@ class AutoSizeText extends StatefulWidget {
   /// The value given to the constructor as textScaleFactor. If null, will
   /// use the [MediaQueryData.textScaleFactor] obtained from the ambient
   /// [MediaQuery], or 1.0 if there is no [MediaQuery] in scope.
-  final double textScaleFactor;
+  final double? textScaleFactor;
 
   /// An optional maximum number of lines for the text to span, wrapping if necessary.
   /// If the text exceeds the given number of lines, it will be resized according
@@ -200,7 +199,7 @@ class AutoSizeText extends StatefulWidget {
   /// an explicit number for its [DefaultTextStyle.maxLines], then the
   /// [DefaultTextStyle] value will take precedence. You can use a [RichText]
   /// widget directly to entirely override the [DefaultTextStyle].
-  final int maxLines;
+  final int? maxLines;
 
   /// An alternative semantics label for this text.
   ///
@@ -212,9 +211,9 @@ class AutoSizeText extends StatefulWidget {
   /// text value:
   ///
   /// ```dart
-  /// Text(r'$$', semanticsLabel: 'Double dollars')
+  /// AutoSizeText(r'$$', semanticsLabel: 'Double dollars')
   /// ```
-  final String semanticsLabel;
+  final String? semanticsLabel;
 
   @override
   _AutoSizeTextState createState() => _AutoSizeTextState();
@@ -222,12 +221,10 @@ class AutoSizeText extends StatefulWidget {
 
 class _AutoSizeTextState extends State<AutoSizeText> {
   @override
-  initState() {
+  void initState() {
     super.initState();
 
-    if (widget.group != null) {
-      widget.group._register(this);
-    }
+    widget.group?._register(this);
   }
 
   @override
@@ -243,90 +240,93 @@ class _AutoSizeTextState extends State<AutoSizeText> {
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(builder: (context, size) {
-      var defaultTextStyle = DefaultTextStyle.of(context);
+      final defaultTextStyle = DefaultTextStyle.of(context);
 
       var style = widget.style;
-      if (widget.style == null || widget.style.inherit) {
+      if (widget.style == null || widget.style!.inherit) {
         style = defaultTextStyle.style.merge(widget.style);
       }
-      if (style.fontSize == null) {
+      if (style!.fontSize == null) {
         style = style.copyWith(fontSize: AutoSizeText._defaultFontSize);
       }
 
-      var maxLines = widget.maxLines ?? defaultTextStyle.maxLines;
+      final maxLines = widget.maxLines ?? defaultTextStyle.maxLines;
 
-      _sanityCheck(style, maxLines);
+      _validateProperties(style, maxLines);
 
-      var result = _calculateFontSize(size, style, maxLines);
-      var fontSize = result[0] as double;
-      var textFits = result[1] as bool;
+      final result = _calculateFontSize(size, style, maxLines);
+      final fontSize = result[0] as double;
+      final textFits = result[1] as bool;
 
       Widget text;
 
       if (widget.group != null) {
-        widget.group._updateFontSize(this, fontSize);
-        text = _buildText(widget.group._fontSize, style, maxLines);
+        widget.group!._updateFontSize(this, fontSize);
+        text = _buildText(widget.group!._fontSize, style, maxLines);
       } else {
         text = _buildText(fontSize, style, maxLines);
       }
 
       if (widget.overflowReplacement != null && !textFits) {
-        return widget.overflowReplacement;
+        return widget.overflowReplacement!;
       } else {
         return text;
       }
     });
   }
 
-  void _sanityCheck(TextStyle style, int maxLines) {
+  void _validateProperties(TextStyle style, int? maxLines) {
     assert(widget.overflow == null || widget.overflowReplacement == null,
-        'Either overflow or overflowReplacement have to be null.');
+        'Either overflow or overflowReplacement must be null.');
     assert(maxLines == null || maxLines > 0,
-        'MaxLines has to be grater than or equal to 1.');
+        'MaxLines must be greater than or equal to 1.');
     assert(widget.key == null || widget.key != widget.textKey,
-        'Key and textKey cannot be the same.');
+        'Key and textKey must not be equal.');
 
     if (widget.presetFontSizes == null) {
-      assert(widget.stepGranularity >= 0.1,
-          'StepGranularity has to be greater than or equal to 0.1. It is not a good idea to resize the font with a higher accuracy.');
+      assert(
+          widget.stepGranularity >= 0.1,
+          'StepGranularity must be greater than or equal to 0.1. It is not a '
+          'good idea to resize the font with a higher accuracy.');
       assert(widget.minFontSize >= 0,
-          'MinFontSize has to be greater than or equal to 0.');
+          'MinFontSize must be greater than or equal to 0.');
       assert(widget.maxFontSize > 0, 'MaxFontSize has to be greater than 0.');
       assert(widget.minFontSize <= widget.maxFontSize,
-          'MinFontSize has to be smaller or equal than maxFontSize.');
+          'MinFontSize must be smaller or equal than maxFontSize.');
       assert(widget.minFontSize / widget.stepGranularity % 1 == 0,
-          'MinFontSize has to be multiples of stepGranularity.');
+          'MinFontSize must be a multiple of stepGranularity.');
       if (widget.maxFontSize != double.infinity) {
         assert(widget.maxFontSize / widget.stepGranularity % 1 == 0,
-            'MaxFontSize has to be multiples of stepGranularity.');
+            'MaxFontSize must be a multiple of stepGranularity.');
       }
     } else {
-      assert(widget.presetFontSizes.isNotEmpty,
-          'PresetFontSizes has to be nonempty.');
+      assert(widget.presetFontSizes!.isNotEmpty,
+          'PresetFontSizes must not be empty.');
     }
   }
 
-  List _calculateFontSize(BoxConstraints size, TextStyle style, int maxLines) {
-    var span = TextSpan(
+  List _calculateFontSize(
+      BoxConstraints size, TextStyle? style, int? maxLines) {
+    final span = TextSpan(
       style: widget.textSpan?.style ?? style,
       text: widget.textSpan?.text ?? widget.data,
       children: widget.textSpan?.children,
       recognizer: widget.textSpan?.recognizer,
     );
 
-    var userScale =
+    final userScale =
         widget.textScaleFactor ?? MediaQuery.textScaleFactorOf(context);
 
     int left;
     int right;
 
-    var presetFontSizes = widget.presetFontSizes?.reversed?.toList();
+    final presetFontSizes = widget.presetFontSizes?.reversed.toList();
     if (presetFontSizes == null) {
-      var defaultFontSize =
-          style.fontSize.clamp(widget.minFontSize, widget.maxFontSize);
-      var defaultScale = defaultFontSize * userScale / style.fontSize;
+      final num defaultFontSize =
+          style!.fontSize!.clamp(widget.minFontSize, widget.maxFontSize);
+      final defaultScale = defaultFontSize * userScale / style.fontSize!;
       if (_checkTextFits(span, defaultScale, maxLines, size)) {
-        return [defaultFontSize * userScale, true];
+        return <Object>[defaultFontSize * userScale, true];
       }
 
       left = (widget.minFontSize / widget.stepGranularity).floor();
@@ -338,12 +338,12 @@ class _AutoSizeTextState extends State<AutoSizeText> {
 
     var lastValueFits = false;
     while (left <= right) {
-      var mid = (left + (right - left) / 2).toInt();
+      final mid = (left + (right - left) / 2).floor();
       double scale;
       if (presetFontSizes == null) {
-        scale = mid * userScale * widget.stepGranularity / style.fontSize;
+        scale = mid * userScale * widget.stepGranularity / style!.fontSize!;
       } else {
-        scale = presetFontSizes[mid] * userScale / style.fontSize;
+        scale = presetFontSizes[mid] * userScale / style!.fontSize!;
       }
       if (_checkTextFits(span, scale, maxLines, size)) {
         left = mid + 1;
@@ -364,56 +364,56 @@ class _AutoSizeTextState extends State<AutoSizeText> {
       fontSize = presetFontSizes[right] * userScale;
     }
 
-    return [fontSize, lastValueFits];
+    return <Object>[fontSize, lastValueFits];
   }
 
   bool _checkTextFits(
-      TextSpan text, double scale, int maxLines, BoxConstraints constraints) {
+      TextSpan text, double scale, int? maxLines, BoxConstraints constraints) {
     if (!widget.wrapWords) {
-      var words = text.toPlainText().split(RegExp('\\s+'));
+      final words = text.toPlainText().split(RegExp('\\s+'));
 
-      var wordWrapTp = TextPainter(
+      final wordWrapTextPainter = TextPainter(
         text: TextSpan(
           style: text.style,
           text: words.join('\n'),
         ),
         textAlign: widget.textAlign ?? TextAlign.left,
         textDirection: widget.textDirection ?? TextDirection.ltr,
-        textScaleFactor: scale ?? 1,
+        textScaleFactor: scale,
         maxLines: words.length,
         locale: widget.locale,
         strutStyle: widget.strutStyle,
       );
 
-      wordWrapTp.layout(maxWidth: constraints.maxWidth);
+      wordWrapTextPainter.layout(maxWidth: constraints.maxWidth);
 
-      if (wordWrapTp.didExceedMaxLines ||
-          wordWrapTp.width > constraints.maxWidth) {
+      if (wordWrapTextPainter.didExceedMaxLines ||
+          wordWrapTextPainter.width > constraints.maxWidth) {
         return false;
       }
     }
 
-    var tp = TextPainter(
+    final textPainter = TextPainter(
       text: text,
       textAlign: widget.textAlign ?? TextAlign.left,
       textDirection: widget.textDirection ?? TextDirection.ltr,
-      textScaleFactor: scale ?? 1,
+      textScaleFactor: scale,
       maxLines: maxLines,
       locale: widget.locale,
       strutStyle: widget.strutStyle,
     );
 
-    tp.layout(maxWidth: constraints.maxWidth);
+    textPainter.layout(maxWidth: constraints.maxWidth);
 
-    return !(tp.didExceedMaxLines ||
-        tp.height > constraints.maxHeight ||
-        tp.width > constraints.maxWidth);
+    return !(textPainter.didExceedMaxLines ||
+        textPainter.height > constraints.maxHeight ||
+        textPainter.width > constraints.maxWidth);
   }
 
-  Widget _buildText(double fontSize, TextStyle style, int maxLines) {
+  Widget _buildText(double fontSize, TextStyle style, int? maxLines) {
     if (widget.data != null) {
       return Text(
-        widget.data,
+        widget.data!,
         key: widget.textKey,
         style: style.copyWith(fontSize: fontSize),
         strutStyle: widget.strutStyle,
@@ -428,7 +428,7 @@ class _AutoSizeTextState extends State<AutoSizeText> {
       );
     } else {
       return Text.rich(
-        widget.textSpan,
+        widget.textSpan!,
         key: widget.textKey,
         style: style,
         strutStyle: widget.strutStyle,
@@ -437,7 +437,7 @@ class _AutoSizeTextState extends State<AutoSizeText> {
         locale: widget.locale,
         softWrap: widget.softWrap,
         overflow: widget.overflow,
-        textScaleFactor: fontSize / style.fontSize,
+        textScaleFactor: fontSize / style.fontSize!,
         maxLines: maxLines,
         semanticsLabel: widget.semanticsLabel,
       );
@@ -451,7 +451,7 @@ class _AutoSizeTextState extends State<AutoSizeText> {
   @override
   void dispose() {
     if (widget.group != null) {
-      widget.group._remove(this);
+      widget.group!._remove(this);
     }
     super.dispose();
   }
