@@ -5,26 +5,25 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-double effectiveFontSize(Text text) {
-  return (text.textScaleFactor ?? 1) * text.style.fontSize;
-}
+double effectiveFontSize(Text text) =>
+    (text.textScaleFactor ?? 1) * text.style!.fontSize!;
 
-bool testIfTextFits(
+bool doesTextFit(
   Text text, [
   double maxWidth = double.infinity,
   double maxHeight = double.infinity,
   bool wrapWords = true,
 ]) {
-  var span = text.textSpan ?? TextSpan(text: text.data, style: text.style);
+  final span = text.textSpan ?? TextSpan(text: text.data, style: text.style);
   var maxLines = text.maxLines;
   if (!wrapWords) {
-    var wordCount = span.toPlainText().split(RegExp('\\s+')).length;
-    maxLines = maxLines.clamp(1, wordCount) as int;
+    final wordCount = span.toPlainText().split(RegExp('\\s+')).length;
+    maxLines = maxLines!.clamp(1, wordCount);
   }
 
-  var tp = TextPainter(
+  final textPainter = TextPainter(
     text: span,
-    textAlign: text.textAlign,
+    textAlign: text.textAlign ?? TextAlign.start,
     textDirection: text.textDirection,
     textScaleFactor: text.textScaleFactor ?? 1,
     maxLines: text.maxLines,
@@ -32,11 +31,11 @@ bool testIfTextFits(
     strutStyle: text.strutStyle,
   );
 
-  tp.layout(maxWidth: maxWidth);
+  textPainter.layout(maxWidth: maxWidth);
 
-  return !(tp.didExceedMaxLines ||
-      tp.height > maxHeight ||
-      tp.width > maxWidth);
+  return !(textPainter.didExceedMaxLines ||
+      textPainter.height > maxHeight ||
+      textPainter.width > maxWidth);
 }
 
 bool prepared = false;
@@ -57,8 +56,8 @@ Future prepareTests(WidgetTester tester) async {
 }
 
 Future pump({
-  @required WidgetTester tester,
-  @required Widget widget,
+  required WidgetTester tester,
+  required Widget widget,
 }) async {
   await tester.pumpWidget(
     Directionality(
@@ -71,25 +70,24 @@ Future pump({
 }
 
 Future<Text> pumpAndGetText({
-  @required WidgetTester tester,
-  @required Widget widget,
+  required WidgetTester tester,
+  required Widget widget,
 }) async {
   await pump(tester: tester, widget: widget);
   return tester.widget<Text>(find.byType(Text));
 }
 
 Future pumpAndExpectFontSize({
-  @required WidgetTester tester,
-  @required double expectedFontSize,
-  @required Widget widget,
+  required WidgetTester tester,
+  required double expectedFontSize,
+  required Widget widget,
 }) async {
-  var text = await pumpAndGetText(tester: tester, widget: widget);
+  final text = await pumpAndGetText(tester: tester, widget: widget);
   expect(effectiveFontSize(text), expectedFontSize);
 }
 
-RichText getRichText(WidgetTester tester) {
-  return tester.widget(find.byType(RichText));
-}
+RichText getRichText(WidgetTester tester) =>
+    tester.widget(find.byType(RichText));
 
 class OverflowNotifier extends StatelessWidget {
   final VoidCallback overflowCallback;
