@@ -7,7 +7,7 @@ import 'max_lines_demo.dart';
 import 'min_font_size_demo.dart';
 import 'overflow_replacement_demo.dart';
 import 'preset_font_sizes_demo.dart';
-import 'step_granulartity.dart';
+import 'step_granularity.dart';
 import 'sync_demo.dart';
 
 void main() {
@@ -25,7 +25,6 @@ class App extends StatelessWidget {
     SystemChrome.setEnabledSystemUIOverlays([]);
 
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
       theme: ThemeData.light(),
       home: DemoApp(),
     );
@@ -33,6 +32,7 @@ class App extends StatelessWidget {
 }
 
 class DemoApp extends StatefulWidget {
+  @override
   _DemoAppState createState() => _DemoAppState();
 }
 
@@ -86,7 +86,7 @@ class _DemoAppState extends State<DemoApp> {
           ),
         ],
         title: Text(
-          'AutoSizeText: ' + demoNames[_selectedDemo],
+          'AutoSizeText: ${demoNames[_selectedDemo]}',
           style: TextStyle(
             color: _selectedColor[500],
             inherit: true,
@@ -96,7 +96,7 @@ class _DemoAppState extends State<DemoApp> {
       body: Container(
         color: _selectedColor[50],
         child: Padding(
-          padding: EdgeInsets.all(15.0),
+          padding: EdgeInsets.all(15),
           child: _buildDemo(),
         ),
       ),
