@@ -23,27 +23,27 @@ class MaxlinesDemo extends StatelessWidget {
                 child: !richText
                     ? Text(
                         input,
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                       )
                     : Text.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                       ),
               ),
             ),
-            SizedBox(width: 10.0),
+            SizedBox(width: 10),
             Expanded(
               child: TextCard(
                 title: 'AutoSizeText',
                 child: !richText
                     ? AutoSizeText(
                         input,
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                         maxLines: 2,
                       )
                     : AutoSizeText.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                         maxLines: 2,
                       ),
               ),
