@@ -13,8 +13,10 @@ class MinFontSizeDemo extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AnimatedInput(
-      text:
-          "This String's size will not be smaller than 20. It will be automatically resized to fit on 4 lines. Otherwise, the String will be ellipsized. Here is some random stuff, just to make sure it is long enough.",
+      text: 'This String\'s size will not be smaller than 20. It will be '
+          'automatically resized to fit on 4 lines. Otherwise, the String will '
+          'be ellipsized. Here is some random stuff, just to make sure it is '
+          'long enough.',
       builder: (input) {
         return Row(
           children: <Widget>[
@@ -24,32 +26,32 @@ class MinFontSizeDemo extends StatelessWidget {
                 child: !richText
                     ? Text(
                         input,
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                         maxLines: 4,
                       )
                     : Text.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 30.0),
+                        style: TextStyle(fontSize: 30),
                         maxLines: 4,
                       ),
               ),
             ),
-            SizedBox(width: 10.0),
+            SizedBox(width: 10),
             Expanded(
               child: TextCard(
                 title: 'AutoSizeText',
                 child: !richText
                     ? AutoSizeText(
                         input,
-                        style: TextStyle(fontSize: 30.0),
-                        minFontSize: 20.0,
+                        style: TextStyle(fontSize: 30),
+                        minFontSize: 20,
                         overflow: TextOverflow.ellipsis,
                         maxLines: 4,
                       )
                     : AutoSizeText.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 30.0),
-                        minFontSize: 20.0,
+                        style: TextStyle(fontSize: 30),
+                        minFontSize: 20,
                         overflow: TextOverflow.ellipsis,
                         maxLines: 4,
                       ),
