@@ -14,7 +14,9 @@ class StepGranularityDemo extends StatelessWidget {
   Widget build(BuildContext context) {
     return AnimatedInput(
       text:
-          "This String changes its size with a stepGranularity of 10. It will be automatically resized to fit on 4 lines. Now the text is so small, you can't even read it...",
+          'This String changes its size with a stepGranularity of 10. It will '
+          'be automatically resized to fit on 4 lines. Now the text is so '
+          'small, you can\'t even read it...',
       builder: (input) {
         return Row(
           children: <Widget>[
@@ -24,34 +26,34 @@ class StepGranularityDemo extends StatelessWidget {
                 child: !richText
                     ? Text(
                         input,
-                        style: TextStyle(fontSize: 40.0),
+                        style: TextStyle(fontSize: 40),
                         maxLines: 4,
                       )
                     : Text.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 40.0),
+                        style: TextStyle(fontSize: 40),
                         maxLines: 4,
                       ),
               ),
             ),
-            SizedBox(width: 10.0),
+            SizedBox(width: 10),
             Expanded(
               child: TextCard(
                 title: 'AutoSizeText',
                 child: !richText
                     ? AutoSizeText(
                         input,
-                        style: TextStyle(fontSize: 40.0),
-                        stepGranularity: 10.0,
-                        minFontSize: 10.0,
+                        style: TextStyle(fontSize: 40),
+                        stepGranularity: 10,
+                        minFontSize: 10,
                         overflow: TextOverflow.ellipsis,
                         maxLines: 4,
                       )
                     : AutoSizeText.rich(
                         spanFromString(input),
-                        style: TextStyle(fontSize: 40.0),
-                        stepGranularity: 10.0,
-                        minFontSize: 10.0,
+                        style: TextStyle(fontSize: 40),
+                        stepGranularity: 10,
+                        minFontSize: 10,
                         overflow: TextOverflow.ellipsis,
                         maxLines: 4,
                       ),
