@@ -12,11 +12,11 @@ class App extends StatelessWidget {
       home: Scaffold(
         body: Center(
           child: SizedBox(
-            width: 200.0,
-            height: 140.0,
+            width: 200,
+            height: 140,
             child: AutoSizeText(
               'This string will be automatically resized to fit in two lines.',
-              style: TextStyle(fontSize: 30.0),
+              style: TextStyle(fontSize: 30),
               maxLines: 2,
             ),
           ),
