@@ -1,4 +1,5 @@
-/// Flutter widget that automatically resizes text to fit perfectly within its bounds.
+/// Flutter widget that automatically resizes text to fit perfectly within its
+/// bounds.
 library auto_size_text;
 
 import 'dart:async';
