@@ -12,19 +12,11 @@ void main() {
     );
   });
 
-  testWidgets('Crash when text null', (tester) async {
-    expect(() => AutoSizeText(null), throwsAssertionError);
-  });
-
   testWidgets('Only text (rich)', (tester) async {
     await pump(
       tester: tester,
       widget: AutoSizeText.rich(TextSpan(text: 'Some Text')),
     );
-  });
-
-  testWidgets('Crash when text null (rich)', (tester) async {
-    expect(() => AutoSizeText.rich(null), throwsAssertionError);
   });
 
   testWidgets('Uses style fontSize', (tester) async {
@@ -50,11 +42,11 @@ void main() {
   });
 
   testWidgets('Respects inherit style', (tester) async {
-    var defaultStyle = TextStyle(
-      fontSize: 20.0,
+    final defaultStyle = TextStyle(
+      fontSize: 20,
       color: Colors.yellow,
     );
-    var text = await pumpAndGetText(
+    final text = await pumpAndGetText(
       tester: tester,
       widget: DefaultTextStyle(
         style: defaultStyle,
@@ -69,7 +61,7 @@ void main() {
     );
     expect(text.style, defaultStyle);
 
-    var richText = getRichText(tester);
+    final richText = getRichText(tester);
     expect(richText.textAlign, TextAlign.right);
     expect(richText.softWrap, false);
     expect(richText.overflow, TextOverflow.ellipsis);
@@ -90,8 +82,8 @@ void main() {
   });
 
   testWidgets('Uses textKey', (tester) async {
-    var textKey = GlobalKey();
-    var text = await pumpAndGetText(
+    final textKey = GlobalKey();
+    final text = await pumpAndGetText(
       tester: tester,
       widget: AutoSizeText(
         'A text with key',
